% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R
\name{acc}
\alias{acc}
\title{Accuracy}
\usage{
acc(y, p, ...)
}
\arguments{
\item{y}{A numerical vector of format n x 1}

\item{p}{A numerical vector of format n x 1}

\item{...}{optional arguments}
}
\value{
A numerical scalar telling the fraction of correctly specified predicitions.
}
\description{
\code{acc} calculates the accuracy, given a binary vector \code{y} and the corresponding
 probability \code{p}. Observations with probability values smaller or equal to 0.5 are classified
as zero, otherwise as 1.
}
\examples{
set.seed(1)
y <- rbinom(100, 1, 0.5)
p <- runif(100, 0, 1)

}
