% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devRes.R
\name{dRes}
\alias{dRes}
\title{Deviance residuals}
\usage{
dRes(y, p, ...)
}
\arguments{
\item{y}{A numerical vector of format n x 1}

\item{p}{A numerical vector of format n x 1}

\item{...}{optional arguments}
}
\value{
A numerical vector containing the deviance residuals.
}
\description{
\code{dRes} calculates the deviance residuals given binary input \code{y} and
the corresponding probability value \code{p}.
}
\examples{
set.seed(1)
y <- rbinom(100, 1, 0.5)
p <- runif(100, 0, 1)
dRes(y, p)

}
