% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher information matrix}
\usage{
fisher(X, p, ...)
}
\arguments{
\item{X}{A numerical matrix of format n x k}

\item{p}{A numerical vector of format n x 1}

\item{...}{optional arguments}
}
\value{
Fisher information matrix.
}
\description{
\code{fisher} calculates the Fisher information matrix, which is 
a symmetric k x k matrix.
}
\examples{
set.seed(1)
X <- replicate(2, rnorm(100))
p <- replicate(100, 0.5)
fisher(X, p)

}
