% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRoc.R
\name{ggRoc}
\alias{ggRoc}
\title{Receiver Operating Characteristic (ROC)}
\usage{
ggRoc(y, p, ...)
}
\arguments{
\item{y}{A numerical vector of format n x 1}

\item{p}{A numerical vector of format n x 1}

\item{...}{optional arguments}
}
\value{
A dataframe containing true postivie and false positive values for each
threshold value.
}
\description{
\code{ggRoc} calculates the confusion matrices for different threshold values in the first
place. In the second place the true positive and false positive rate will be calculated
for each threshold value.
}
\examples{
set.seed(1)
y <- rbinom(100, 1, 0.5)
p <- runif(100, 0, 1)
ggRoc(y, p)

}
