% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iQuantile.R
\name{iQuantile}
\alias{iQuantile}
\title{iQuantile}
\format{\describe{
\item{x}{
The function has two arguments: a  list and numnber of breaks.}
}}
\usage{
iQuantile(x, breaks=15)
}
\arguments{
\item{x}{list}

\item{breaks}{number of breaks}
}
\value{
numeric
}
\description{
Produces indices in x  of percentile steps by 1/breaks and respective percentiles.
}
\details{
iQuantile is used internally by the  hlGOF.test.
}
\examples{
library(LOGIT)
mod <- rnorm(100,0,1)
iQuantile(mod)
}
\author{
Joseph M. Hilbe, Arizona State University.
}
\references{
Hilbe, J. M. (2015), Practical Guide to Logistic Regression, Chapman & Hall/CRC.

Hilbe, J. M. (2009), Logistic Regression Models, Chapman & Hall/CRC.
}
\keyword{models}

