% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lNull.R
\name{lNull}
\alias{lNull}
\title{Log likelihood of intercept model (Null model)}
\usage{
lNull(y, ...)
}
\arguments{
\item{y}{A numerical vector of format n x 1}

\item{...}{optional arguments}
}
\value{
A numerical scalar containing the log likelihood of the Null model.
}
\description{
\code{lNull} calculates the log of the likelihood for a logistic regression model containing
only the intercept.
}
\examples{
set.seed(1)
y <- rbinom(100, 1, 0.5)

}
