% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logisitc function}
\usage{
logistic(x, ...)
}
\arguments{
\item{x}{A numeric vector or scalar}

\item{...}{optional arguments}
}
\value{
A numeric vector or scalar after applying logistic function on a original vector.
}
\description{
\code{logistic} transforms a domain of real numbers from negative infity to positive infinity
to real numbers lying between zero and one.
}
\examples{
a <- -5:5
logistic(a)

}
