% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{logit.default}
\alias{logit.formula}
\title{Logistic Regression}
\usage{
logit(x, ...)

\method{logit}{default}(x, y, ...)

\method{logit}{formula}(formula, data = list(), ...)
}
\arguments{
\item{x}{An object of any class.}

\item{...}{optional arguments}

\item{y}{A binary vector of format n x 1}

\item{formula}{way of extracting formulae which have been included in other objects}

\item{data}{a list including all relevant variables}
}
\value{
a list of class "logit" containing a set of model characteristica:
\itemize{
\item X
\item y
\item Log Likelihood for each iteration
\item Convergence
\item estimated parameter vector
\item Estimated probability for y=1
\item Deviance  
\item Deviance residuals 
\item Pearson resiudals
\item AIC for each iteration
\item Accuracy
\item Variance Covariance Matrix
\item McFadden R squared
\item Degrees of freedom
\item Degrees of freedom of Null model
}
}
\description{
\code{logit} performs a logistic regression given a binary 
response variable and a set of regressors.
}
\examples{
logit(Survived ~ ., data = myData)

}
