% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLike.R
\name{logLike}
\alias{logLike}
\title{Log likelihood function}
\usage{
logLike(X, y, beta, ...)
}
\arguments{
\item{X}{A numerical matrix of format n x k}

\item{y}{A binary numerical vector of format n x 1}

\item{beta}{A numerical vector of format k x 1}

\item{...}{optional arguments}
}
\value{
A scalar containing the log likelihood value.
}
\description{
\code{logLike} calculates the log of the likelihood function.
The likelihood whether a specific event occurs (\code{y}=1) is given by a probabilty value which
follows from the logistic function given some observed data \code{X}.
}
\examples{
set.seed(1)
X <- replicate(2, rnorm(100))
y <- rbinom(100, 1, 0.5)
beta <- c(0, 0)
logLike(X, y, beta)

}
