% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.logit.R
\name{plot.logit}
\alias{plot.logit}
\title{Plot method for objects of class "logit"}
\usage{
\method{plot}{logit}(x, ...)
}
\arguments{
\item{x}{An object of class "logit".}

\item{...}{optional arguments}
}
\value{
Three plots for objects of class "logit" will be generated:
\enumerate{
\item Linear predictons against deviance residuals 
\item A Normal QQ Plot
\item A Confusion Matrix
}
}
\description{
\code{plot} method for objects of class "logit".
}
\examples{
mod <- logit(Survived ~ ., data = myData)
plot(mod)

}
