% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.roc.R
\name{plot.roc}
\alias{plot.roc}
\title{Plot ROC curve  for objects of class "roc"}
\usage{
\method{plot}{roc}(x, ...)
}
\arguments{
\item{x}{A list of class "roc".}

\item{...}{optional arguments}
}
\value{
a plot containing a roc curve
}
\description{
\code{plot} method for objects of class "roc".
}
\examples{
mod <- logit(Survived ~ ., data = myData)
obj <- ggRoc(mod$y, mod$prob)
plot(obj)

}
