% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.logit.R
\name{predict.logit}
\alias{predict.logit}
\title{Predict method for objects of class "logit"}
\usage{
\method{predict}{logit}(object, t = 0.5, ...)
}
\arguments{
\item{object}{An object of class "logit"}

\item{t}{A numerical scalar taking a value between 0 and 1}

\item{...}{optional arguments}
}
\value{
A matrix having three columns : 
\itemize{
\item Linear Predictor
\item Probability
\item Predicted class (zero or one)
}
}
\description{
\code{predict} method for objects of class "logit".
}
\details{
Observations with predicted probability smaller or equal to \code{t} will be
assigned to category zero. In reverse observations with predicted probability \code{p} larger 
than \code{t} will be be assigned to category one.
}
\examples{
mod <- logit(Survived ~ ., data = myData)
predict(mod)

}
