% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearRes.R
\name{pRes}
\alias{pRes}
\title{Pearson residuals}
\usage{
pRes(y, p, ...)
}
\arguments{
\item{y}{A numerical vector of format n x 1}

\item{p}{A numerical vector of format n x 1}

\item{...}{optional arguments}
}
\value{
A numerical vector containing the pearson residuals.
}
\description{
\code{pRes} calculates the pearson residuals given binary input \code{y} and
the corresponding probability value \code{p}.
}
\examples{
set.seed(1)
y <- rbinom(100, 1, 0.5)
p <- runif(100, 0, 1)

}
