% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.logit.R
\name{print.summary.logit}
\alias{print.summary.logit}
\title{Print method for objects of class "summary.logit"}
\usage{
\method{print}{summary.logit}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.logit".}

\item{...}{optional arguments}
}
\value{
A summary table for objects of class "summary.logit".
}
\description{
Method to determine a \code{summary} table for objects of class "\code{logit}".
Gives information about the estimated parameters, their significance and further 
model characeristica.
}
\examples{
mod <- logit(Survived ~ ., data = myData)
summary(mod)
}
