% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score vector}
\usage{
score(X, y, beta, ...)
}
\arguments{
\item{X}{A numerical matrix of format n x k}

\item{y}{A numerical vector of format n x 1}

\item{beta}{A numerical vector of format k x 1}

\item{...}{optional arguments}
}
\value{
A numerical vector containing the score.
}
\description{
\code{score} calculates the first derivative of the log likelihood function with respect to beta, given
some observed data \code{X} and \code{y}.
}
\examples{
set.seed(1)
X <- replicate(2, rnorm(100))
y <- rbinom(100, 1, 0.5)
beta <- c(0, 0)
score(X, y, beta)

}
