% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.logit.R
\name{summary.logit}
\alias{summary.logit}
\title{Summary method for objects of class "logit"}
\usage{
\method{summary}{logit}(object, ...)
}
\arguments{
\item{object}{An object of class "logit".}

\item{...}{optional arguments}
}
\value{
A list of class "summary.logit". It is a list containing the 
following elements:
\itemize{
\item \emph{coefficients}: Estimated coefficients
\item \emph{Std. Err}: Standard Erros of estimated coefficients
\item \emph{z value}: Test statisitic of estimated coefficients
\item \emph{{P(>|z|)}}: P-Values for estimated coefficients 
\item \emph{call}: Function call
\item \emph{NullDeviance}: Deviance of Intercept model
\item \emph{NullDf}: Degrees of freedom for Null model 
\item \emph{NullDf}: Degrees of freedom for Null model 
\item \emph{mcFad}: Mc Fadden R squared
\item \emph{y}: binary response variable
\item \emph{X}: Set of regressors
\item \emph{Accuracy}: Accuracy of estimated model
\item \emph{AIC}: AIC of estimated model
\item \emph{iterations}: Number of Fisher scoring iterations needed to converge
\item \emph{residuals}: Deviance residuals of estimated model
\item \emph{Pearson}: Pearson residuals of estimated model
}
}
\description{
Method to get a \code{summary} list for objects of class "logit".
Gives information about the estimated parameters, their significance and further 
model characeristica.
}
\examples{
mod <- logit(Survived ~ ., data = myData)
str(mod) # list 
# To access the different elements of the summary the "\\code$" can used, 
# e.g. taking a look at the coefficients:
mod$coefficients

}
