% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/confusion_stat.R
\name{confusion_stat}
\alias{confusion_stat}
\title{Display confusion or classification matrix following logistic regression using glm with binomial family.}
\format{\describe{
\item{x}{
The function has two arguments: predicted values, response values}
}}
\usage{
confusion_stat(<predicted statistic>, <observed response>)
}
\arguments{
\item{pred}{}

\item{obs}{}
}
\value{
confusion matrix
}
\description{
Provides a confusion matrix of classification statistics following logistic regression.
}
\note{
confusion_stat() must be loaded into memory in order to be effectve. As a function in LOGIT,
it is immediately available to a user.
}
\examples{
library(MASS)
 library(LOGIT)
 data(medpar)
 mylogit <- glm( died ~  los + white + hmo, family=binomial, data=medpar)
 mu <- predict(mylogit, type="response")
 confusion_stat(mu, medpar$died)
}
\author{
Rafael de Souza, ELTE  University,  and Joseph M. Hilbe, Arizona State University
}
\references{
Hilbe, Joseph M. (2016), Practical Guide to Logistic Regression, Chapman & Hall/CRC.
Hilbe, Joseph M. (2009), Logistic Regression Models, Chapman & Hall/CRC.
}
\seealso{
\code{\link{glm}}
}
\keyword{models}

